% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_frequency.R
\name{amino_acid_frequency}
\alias{amino_acid_frequency}
\alias{aa_freq}
\title{Calculate the frequency of amino acids in a protein sequence}
\usage{
amino_acid_frequency(protein_seq)

aa_freq(protein_seq)
}
\arguments{
\item{protein_seq}{A character string representing the protein sequence}
}
\value{
A data frame with two columns: Amino_Acid and Count, where Amino_Acid
is the name of the amino acid and Count is the number of times it occurs
in the protein sequence.
}
\description{
This function takes a protein sequence as input and returns a data frame
summarizing the frequency of each amino acid in the sequence.
}
\examples{
amino_acid_frequency("MSRSLLLRFLLFLLLLPPLP")

}
