% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_library.R
\name{amino_acid_library}
\alias{amino_acid_library}
\alias{aa_library}
\title{Get a vector of amino acid codes, names, or color palette}
\usage{
amino_acid_library(names)

aa_library(names)
}
\arguments{
\item{names}{character vector. Names of the requested amino acid library components.
Possible values are "aa_codes", "amino_acid_codes", "ac", "aa_names", "amino_acid_names", "an",
"aa_colors", "amino_acid_colors", "acp".}
}
\value{
A character vector containing amino acid codes, names, or color palette.
}
\description{
Get a vector of amino acid codes, names, or color palette
}
\details{
This function returns a vector of amino acid codes, names, or a color palette
for visualizing amino acids. The color palette consists of 21 colors, each corresponding
to an amino acid or stop codon. The function can be used with three different argument values:

- "aa_codes", "amino_acid_codes", "ac": returns a character vector of amino acid codes.
- "aa_names", "amino_acid_names", "an": returns a character vector of amino acid names.
- "aa_colors", "amino_acid_colors", "acp": returns a character vector of color codes
corresponding to amino acids and stop codons.
}
\examples{
amino_acid_library("aa_names")
amino_acid_library("amino_acid_codes")
amino_acid_library("amino_acid_colors")

}
