% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as_baseq_dna}
\alias{as_baseq_dna}
\title{S3 DNA Class}
\usage{
as_baseq_dna(s)
}
\arguments{
\item{s}{A character string containing the sequence}
}
\value{
A baseq_dna object
}
\description{
Creates an S3 object of class baseq_dna.
}
