% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as_baseq_rna}
\alias{as_baseq_rna}
\title{S3 RNA Class}
\usage{
as_baseq_rna(s)
}
\arguments{
\item{s}{A character string containing the sequence}
}
\value{
A baseq_rna object
}
\description{
Creates an S3 object of class baseq_rna.
}
