% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as_Biostrings}
\alias{as_Biostrings}
\title{Bioconductor Bridge}
\usage{
as_Biostrings(s)
}
\arguments{
\item{s}{A character vector or list of sequences}
}
\value{
A DNAStringSet object
}
\description{
Converts baseq sequences to Biostrings format.
}
