% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_calculate_extinction_coefficient.R
\name{bulk_calculate_extinction_coefficient}
\alias{bulk_calculate_extinction_coefficient}
\title{Calculate Extinction Coefficient for Proteins in a Data Frame}
\usage{
bulk_calculate_extinction_coefficient(protein_df)
}
\arguments{
\item{protein_df}{A data frame containing a column named "Sequence" with the protein sequences to calculate extinction coefficients for.}
}
\value{
A data frame with a new column named "Extinction_Coefficient" containing the extinction coefficients for each protein sequence.
}
\description{
This function calculates the extinction coefficient for each protein sequence in a data frame and appends the result as a new column in the same data frame.
}
\examples{
protein <- generate_multi_protein_sequences_as_df(10,50)
bulk_calculate_extinction_coefficient(protein)

}
