% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlab.R
\name{calculate_charge}
\alias{calculate_charge}
\title{Protein Net Charge}
\usage{
calculate_charge(s, ph = 7.4)
}
\arguments{
\item{s}{A character string containing the protein sequence}

\item{ph}{Numeric pH value (default: 7.4)}
}
\value{
Numeric net charge
}
\description{
Calculates the net electrical charge of a protein at a given pH.
}
