% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calculate_codon_usage}
\alias{calculate_codon_usage}
\title{Codon Usage RSCU}
\usage{
calculate_codon_usage(s)
}
\arguments{
\item{s}{A character string containing the coding DNA sequence}
}
\value{
A dataframe with codon statistics
}
\description{
Calculates Relative Synonymous Codon Usage (RSCU).
}
\examples{
data(sars_fragment)
calculate_codon_usage(sars_fragment)
}
