% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_extinction_coefficient.R
\name{calculate_extinction_coefficient}
\alias{calculate_extinction_coefficient}
\title{Calculate extinction coefficient}
\usage{
calculate_extinction_coefficient(protein_sequence)
}
\arguments{
\item{protein_sequence}{A character vector representing the protein sequence.}
}
\value{
The extinction coefficient of the protein sequence at 280 nm.
}
\description{
This function calculates the extinction coefficient of a protein sequence at 280 nm based on the molar extinction coefficients of its constituent amino acids.
}
\examples{
calculate_extinction_coefficient("MKTIIALSYIFCLGLAGSTNVTEAITCQQLYQFVSGNLEKI")

}
