% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calculate_identity}
\alias{calculate_identity}
\title{Sequence Identity}
\usage{
calculate_identity(s1, s2)
}
\arguments{
\item{s1}{First sequence}

\item{s2}{Second sequence}
}
\value{
A list with Identity percentage and Hamming Distance
}
\description{
Compares two sequences of equal length.
}
\examples{
calculate_identity("ATGC", "ATGG")
}
