% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_molecular_weight_dna.R
\name{calculate_molecular_weight_dna}
\alias{calculate_molecular_weight_dna}
\alias{mw_dna}
\title{Calculate molecular weight of a DNA sequence}
\usage{
calculate_molecular_weight_dna(dna_seq)

mw_dna(dna_seq)
}
\arguments{
\item{dna_seq}{A DNA sequence as a character string}
}
\value{
The molecular weight of the DNA sequence in g/mol
}
\description{
This function calculates the molecular weight of a given DNA sequence.
It uses a table of molecular weights of nucleotides to calculate the molecular weight of the sequence.
}
\examples{
calculate_molecular_weight_dna("ATGCGCTA")
# [1] 1423.4
}
