% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_molecular_weight_protein.R
\name{calculate_molecular_weight_protein}
\alias{calculate_molecular_weight_protein}
\alias{mw_protein}
\title{Calculate Molecular Weight of a Protein Sequence}
\usage{
calculate_molecular_weight_protein(protein_seq)

mw_protein(protein_seq)
}
\arguments{
\item{protein_seq}{A character string representing the protein sequence. It should consist only of the standard amino acid symbols (A, C, D, E, F, G, H, I, K, L, M, N, P, Q, R, S, T, V, W, Y) in uppercase.}
}
\value{
A numeric value representing the molecular weight of the protein sequence in g/mol.
}
\description{
This function calculates the molecular weight of a given protein sequence based on the molecular weights of its constituent amino acids.
}
\examples{
calculate_molecular_weight_protein("MKGHHLMLLGLLGLVLALGGSSSAQAKL")

}
