% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_molecular_weight_rna.R
\name{calculate_molecular_weight_rna}
\alias{calculate_molecular_weight_rna}
\alias{mw_rna}
\title{Calculate the molecular weight of an RNA sequence}
\usage{
calculate_molecular_weight_rna(rna_seq)

mw_rna(rna_seq)
}
\arguments{
\item{rna_seq}{A character string representing the RNA sequence}
}
\value{
A numeric value representing the molecular weight of the RNA sequence in g/mol
}
\description{
This function calculates the molecular weight of an RNA sequence using the following formula:
}
\details{
\code{Molecular weight = sum of (number of nucleotides of each type * molecular weight of each nucleotide)}
}
\examples{
calculate_molecular_weight_rna("AGCUUAGCUA")

}
