% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlab.R
\name{calculate_mw}
\alias{calculate_mw}
\title{Protein MW}
\usage{
calculate_mw(s)
}
\arguments{
\item{s}{A character string containing the protein sequence}
}
\value{
Numeric molecular weight in Daltons
}
\description{
Calculates the molecular weight of a protein sequence.
}
