% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlab.R
\name{calculate_pi}
\alias{calculate_pi}
\title{Protein pI}
\usage{
calculate_pi(s)
}
\arguments{
\item{s}{A character string containing the protein sequence}
}
\value{
Numeric pI value
}
\description{
Estimates the isoelectric point of a protein sequence.
}
