% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlab.R
\name{calculate_tm}
\alias{calculate_tm}
\title{Primer Tm}
\usage{
calculate_tm(s, salt = 50)
}
\arguments{
\item{s}{A character string containing the sequence}

\item{salt}{Numeric salt concentration in mM (default: 50)}
}
\value{
Numeric Tm in Celsius
}
\description{
Calculates the melting temperature of a primer sequence.
}
