% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_DNA_sequence.R
\name{clean_DNA_sequence}
\alias{clean_DNA_sequence}
\title{Clean DNA sequence}
\usage{
clean_DNA_sequence(sequence)
}
\arguments{
\item{sequence}{DNA sequence to be cleaned}
}
\value{
Cleaned DNA sequence
}
\description{
This function takes a DNA sequence as input and removes any characters other than A, C, G, and T.
}
\examples{
clean_DNA_sequence("ATGTCGTAGCTAGCTN")
# Output: "ATGTCGTAGCTAGCT"

}
