% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{clean_file}
\alias{clean_file}
\title{Batch File Cleaner}
\usage{
clean_file(input_file, type = "auto", output_dir = "")
}
\arguments{
\item{input_file}{Path to input file}

\item{type}{Sequence type ("DNA", "RNA", or "auto")}

\item{output_dir}{Optional output directory}
}
\value{
Path to the cleaned file
}
\description{
Cleans all sequences in a FASTA or FASTQ file.
}
