% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_protein_file.R
\name{clean_protein_file}
\alias{clean_protein_file}
\title{Clean protein sequences from a file and save to a new file}
\usage{
clean_protein_file(input_file, output_dir = "")
}
\arguments{
\item{input_file}{character string specifying the path to the input file of protein sequences in FASTA format.}

\item{output_dir}{character string specifying the directory where the cleaned sequences file should be saved. Defaults to the same directory as the input file.}
}
\value{
A character string specifying the path to the file where the cleaned sequences have been saved.
}
\description{
This function reads in a file of protein sequences in FASTA format, removes any non-amino acid characters, and saves the cleaned sequences to a new file.
}
\examples{
# clean_protein_file("path/to/input/file.fasta", output_dir = "path/to/output/directory")

}
