% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_protein_sequence.R
\name{clean_protein_sequence}
\alias{clean_protein_sequence}
\alias{clean_prot}
\title{Clean a protein sequence by removing non-standard amino acid codes}
\usage{
clean_protein_sequence(sequence)

clean_prot(sequence)
}
\arguments{
\item{sequence}{A character string representing a protein sequence.}
}
\value{
A character string representing the cleaned protein sequence.
}
\description{
This function takes a protein sequence as input and removes any characters
that are not standard amino acid codes. It then returns the cleaned sequence.
}
\examples{
clean_protein_sequence("MAEGEITTFTALTEK$^3556FNLPPGNYKKPKLLCSNGGHDGTVDGTRDR")

}
