% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_RNA_sequence.R
\name{clean_RNA_sequence}
\alias{clean_RNA_sequence}
\title{Clean RNA sequence}
\usage{
clean_RNA_sequence(sequence)
}
\arguments{
\item{sequence}{RNA sequence to be cleaned}
}
\value{
Cleaned RNA sequence
}
\description{
This function takes a RNA sequence as input and removes any characters other than A, C, G, and T.
}
\examples{
clean_RNA_sequence("AUGUCGTAGCTAGCTN")
# Output: "AUGUCGAGCAGC"

}
