% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{clean_seq}
\alias{clean_seq}
\title{Universal Sequence Cleaner}
\usage{
clean_seq(sequence, type = "auto")
}
\arguments{
\item{sequence}{A character string containing the sequence}

\item{type}{A string "DNA", "RNA", or "auto"}
}
\value{
A character string of the cleaned sequence
}
\description{
Removes non-standard characters from DNA or RNA sequences.
}
