% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_frequency.R
\name{codon_frequency}
\alias{codon_frequency}
\alias{codon_freq}
\title{Calculate the frequency of each codon in a DNA sequence}
\usage{
codon_frequency(dna_seq)

codon_freq(dna_seq)
}
\arguments{
\item{dna_seq}{A character string representing a DNA sequence.}
}
\value{
A data frame with two columns: Codon (the name of each codon) and Count
(the frequency of each codon in the input DNA sequence).
}
\description{
This function takes a DNA sequence and calculates the frequency of each
codon in the sequence. The function returns a data frame with the name of
each codon and its frequency in the sequence.
}
\examples{
codon_frequency("ATGCGCTGCTAG")

}
