% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_frequency_plot.R
\name{codon_frequency_plot}
\alias{codon_frequency_plot}
\alias{cf_plot}
\title{Plot the codon frequency of a DNA sequence}
\usage{
codon_frequency_plot(dna_seq)

cf_plot(dna_seq)
}
\arguments{
\item{dna_seq}{A character string representing a DNA sequence.}
}
\value{
A bar plot representing the frequency of each codon in the input DNA sequence.
}
\description{
This function creates a bar plot representing the frequency of each codon in a given DNA sequence.
}
\examples{
#codon_frequency_plot("ATGCGCTGCTAG")

}
