% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_library.R
\name{codon_library}
\alias{codon_library}
\title{Codon Library}
\usage{
codon_library(names)
}
\arguments{
\item{names}{A character vector indicating the name of the output. Can be one of "codon_codes", "amino_acid_triplets",
"aa_triplets", "cc" for codon codes, or "codon_names","aa_triplet_names","amino_acid_triplet_names", "aa_names", "cn" for codon names.}
}
\value{
A vector of codon codes or codon names based on the input name.
}
\description{
This function returns a vector of codon codes or codon names based on the input name.
}
\details{
The function contains a fixed vector of codon codes and codon names for each amino acid. The function returns
either a vector of codon codes or codon names based on the input name. If an invalid name is passed as input, an error
message is printed and the function stops.
}
\examples{
codon_library("codon_codes")
codon_library("codon_names")
codon_library("cc")
codon_library("cn")

}
