% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{count_bases}
\alias{count_bases}
\title{Count Bases}
\usage{
count_bases(s)
}
\arguments{
\item{s}{A character string containing the sequence}
}
\value{
A table object with base counts
}
\description{
Returns a frequency table of the bases in a sequence.
}
\examples{
data(sars_fragment)
count_bases(sars_fragment)
}
