% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{count_kmers}
\alias{count_kmers}
\title{K-mer Counting}
\usage{
count_kmers(s, k = 3)
}
\arguments{
\item{s}{A character string containing the sequence}

\item{k}{Integer length of k-mer}
}
\value{
A table of k-mer counts
}
\description{
Counts all possible substrings of length k.
}
\examples{
data(sars_fragment)
count_kmers(sars_fragment, k = 3)
}
