% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{count_pattern}
\alias{count_pattern}
\title{Count Pattern}
\usage{
count_pattern(s, p)
}
\arguments{
\item{s}{A character string containing the sequence}

\item{p}{A character string containing the pattern to count}
}
\value{
Integer count of occurrences
}
\description{
Counts the occurrences of a specific pattern in a sequence.
}
\examples{
data(sars_fragment)
count_pattern(sars_fragment, "ATTA")
}
