% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_seq_pattern.R
\name{count_seq_pattern}
\alias{count_seq_pattern}
\title{Count frequency of a pattern in a sequence}
\usage{
count_seq_pattern(seq, pattern)
}
\arguments{
\item{seq}{A character vector representing the sequence to count the pattern in.}

\item{pattern}{A character string representing the pattern to count in the sequence.}
}
\value{
An integer representing the count of the pattern in the sequence.
}
\description{
This function counts the frequency of a specific character or pattern in a given sequence.
}
\examples{
seq <- "ATGGTGCTCCGTGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCGCTACGTAG"
count_seq_pattern(seq, "CG")
# [1] 31

}
