% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{dna_to_protein}
\alias{dna_to_protein}
\title{Translate DNA to Protein}
\usage{
dna_to_protein(s, table = 1)
}
\arguments{
\item{s}{A character string containing the DNA sequence}

\item{table}{Integer indicating the NCBI genetic code table (default: 1)}
}
\value{
A list of translated protein sequences
}
\description{
Translates a DNA sequence into protein in all 6 reading frames.
}
