% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{DNA to RNA}
\usage{
dna_to_rna(s)
}
\arguments{
\item{s}{A character string containing the DNA sequence}
}
\value{
A character string of the RNA sequence
}
\description{
Transcribes a DNA sequence into RNA.
}
