% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_stats.R
\name{fasta.stats}
\alias{fasta.stats}
\alias{fasta_stats}
\title{Calculate basic statistics for multi-fasta sequences}
\usage{
fasta.stats(file_path)

fasta_stats(file_path)
}
\arguments{
\item{file_path}{A character string specifying the path to the multi-fasta file to be analyzed.}
}
\value{
A list of named lists containing sequence statistics for each sequence in the multi-fasta file.
}
\description{
This function reads in a multi-fasta file and calculates sequence statistics such as sequence length, GC content, and base counts for each sequence.
}
\details{
This function reads in a multi-fasta file specified by {file_path} and calculates various statistics for each sequence. The statistics include the length of the sequence, GC content, and counts of each base type (A, C, G, T). The results are returned as a list of named lists, with each element of the list corresponding to a single sequence in the multi-fasta file.
}
\examples{
# fasta_stats("example.fasta")

}
