% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_to_df_stats.R
\name{fasta_to_df_stats}
\alias{fasta_to_df_stats}
\title{Calculate statistics from a FASTA dataframe}
\usage{
fasta_to_df_stats(fasta_df)
}
\arguments{
\item{fasta_df}{A dataframe in which the first column contains the sequence headers and the second column contains the sequences in FASTA format.}
}
\value{
This function does not return anything. Instead, it prints the statistics to the console.
}
\description{
This function takes a FASTA dataframe as input and calculates statistics for each sequence as well as overall statistics for the entire dataset. The output is printed to the console.
}
\examples{
fasta_df <- data.frame(Header = c(">seq1", ">seq2"), Sequence = c("ATCG", "ATCGAT"))
fasta_to_df_stats(fasta_df)

}
