% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fastq_to_fasta}
\alias{fastq_to_fasta}
\title{Convert FASTQ to FASTA}
\usage{
fastq_to_fasta(fastq_file)
}
\arguments{
\item{fastq_file}{Path to input FASTQ}
}
\value{
Path to output FASTA
}
\description{
Converts a FASTQ file to FASTA format.
}
