% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter_fastq_quality}
\alias{filter_fastq_quality}
\title{Quality Filter FASTQ}
\usage{
filter_fastq_quality(
  input_file,
  output_file,
  min_avg_quality = 20,
  phred_offset = 33
)
}
\arguments{
\item{input_file}{Path to input FASTQ}

\item{output_file}{Path to output FASTQ}

\item{min_avg_quality}{Minimum average Phred score (default: 20)}

\item{phred_offset}{Phred offset (default: 33)}
}
\description{
Filters FASTQ reads based on average quality score.
}
