% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{find_cpg_islands}
\alias{find_cpg_islands}
\title{CpG Island Detection}
\usage{
find_cpg_islands(s, window = 200)
}
\arguments{
\item{s}{A character string containing the DNA sequence}

\item{window}{Sliding window size (default: 200)}
}
\value{
A dataframe with start and end positions
}
\description{
Identifies candidate CpG islands in a DNA sequence.
}
