% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{find_longest_orf}
\alias{find_longest_orf}
\title{Find Longest ORF}
\usage{
find_longest_orf(s)
}
\arguments{
\item{s}{A character string containing the DNA sequence}
}
\value{
A character string of the longest translated protein sequence
}
\description{
Scans a DNA sequence in all 6 reading frames to find the longest open reading frame.
}
