% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{gc_content}
\alias{gc_content}
\title{GC Content}
\usage{
gc_content(s)
}
\arguments{
\item{s}{A character string containing the sequence}
}
\value{
Numeric percentage of GC content
}
\description{
Calculates the percentage of G and C bases in a DNA sequence.
}
\examples{
data(sars_fragment)
gc_content(sars_fragment)
}
