% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_content_file.R
\name{gc_content_file}
\alias{gc_content_file}
\title{GC content of sequences in a multi FASTA file}
\usage{
gc_content_file(input_file)
}
\arguments{
\item{input_file}{A string indicating the path and name of the input multi-FASTA file}
}
\description{
Function to calculate GC content of sequences in a multi FASTA file
and write the results to a new FASTA file
}
\examples{
#sample_file_path <- system.file("extdata", "sample_fa.fasta", package = "baseq")
#clean_DNA_file(sample_file_path)

}
