% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_illumina_fastq_sequences.R
\name{generate_illumina_fastq_sequences}
\alias{generate_illumina_fastq_sequences}
\title{Generate Illumina Fastq Sequences}
\usage{
generate_illumina_fastq_sequences(num_seq, seq_length)
}
\arguments{
\item{num_seq}{An integer specifying the number of sequences to generate.}

\item{seq_length}{An integer specifying the length of each sequence.}
}
\value{
A character vector containing the generated sequences in fastq format.
}
\description{
This function generates artificial Illumina sequencing reads in fastq format.
}
\examples{
generate_illumina_fastq_sequences(3, 50)

}
