% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multi_dna_seq.R
\name{generate_multi_dna_sequences}
\alias{generate_multi_dna_sequences}
\title{Generate multiple random DNA sequences}
\usage{
generate_multi_dna_sequences(num_seq, seq_length)
}
\arguments{
\item{num_seq}{The number of sequences to generate}

\item{seq_length}{The length of each sequence}
}
\value{
A character vector of DNA sequences in FASTA format
}
\description{
This function generates a specified number of random DNA sequences of a specified length.
}
\examples{
generate_multi_dna_sequences(3, 10)

}
