% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multi_dna_sequences_as_list.R
\name{generate_multi_dna_sequences_as_list}
\alias{generate_multi_dna_sequences_as_list}
\title{Generate multiple DNA sequences as a list}
\usage{
generate_multi_dna_sequences_as_list(num_seq, seq_length)
}
\arguments{
\item{num_seq}{The number of DNA sequences to generate.}

\item{seq_length}{The length of each DNA sequence.}
}
\value{
A list of DNA sequences.
}
\description{
Generate multiple DNA sequences as a list
}
\examples{
generate_multi_dna_sequences_as_list(10, 50)

}
