% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multi_fastq_sequences.R
\name{generate_multi_fastq_sequences}
\alias{generate_multi_fastq_sequences}
\title{Generate multiple FASTQ sequences}
\usage{
generate_multi_fastq_sequences(num_seq, seq_length)
}
\arguments{
\item{num_seq}{Number of sequences to generate.}

\item{seq_length}{Length of each sequence.}
}
\value{
A character vector of generated FASTQ sequences.
}
\description{
This function generates a specified number of FASTQ sequences with random nucleotides and quality scores.
}
\examples{
generate_multi_fastq_sequences(5, 50)

}
