% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multi_fastq_sequences_as_df.R
\name{generate_multi_fastq_sequences_as_df}
\alias{generate_multi_fastq_sequences_as_df}
\title{Generate multiple FASTQ sequences as a data frame}
\usage{
generate_multi_fastq_sequences_as_df(num_seq, seq_length)
}
\arguments{
\item{num_seq}{Number of sequences to generate.}

\item{seq_length}{Length of the sequences to generate.}
}
\value{
A data frame containing the generated FASTQ sequences.
}
\description{
This function generates a specified number of FASTQ sequences with a specified sequence length and returns them as a data frame.
}
\examples{
generate_multi_fastq_sequences_as_df(5, 20)

}
