% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multi_fastq_sequences_as_list.R
\name{generate_multi_fastq_sequences_as_list}
\alias{generate_multi_fastq_sequences_as_list}
\title{Generate Multiple FASTQ Sequences as List}
\usage{
generate_multi_fastq_sequences_as_list(num_seq, seq_length)
}
\arguments{
\item{num_seq}{An integer indicating the number of sequences to be generated.}

\item{seq_length}{An integer indicating the length of each sequence to be generated.}
}
\value{
A list of multiple FASTQ sequences. Each sequence contains two elements: "Sequence" and "QualityScore".
}
\description{
This function generates a list of multiple FASTQ sequences. Each sequence in the list contains two elements: the DNA sequence and the quality score.
}
\examples{
generate_multi_fastq_sequences_as_list(5, 20)

}
