% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multi_protein_seq.R
\name{generate_multi_protein_sequences}
\alias{generate_multi_protein_sequences}
\title{Generate Multiple Protein Sequences}
\usage{
generate_multi_protein_sequences(num_seq, seq_length)
}
\arguments{
\item{num_seq}{An integer specifying the number of protein sequences to generate.}

\item{seq_length}{An integer specifying the length of each protein sequence.}
}
\value{
This function returns a character vector containing the generated protein sequences.
}
\description{
This function generates multiple protein sequences of specified length.
}
\examples{
generate_multi_protein_sequences(3, 10)

}
