% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multi_protein_sequences_as_df.R
\name{generate_multi_protein_sequences_as_df}
\alias{generate_multi_protein_sequences_as_df}
\title{Generate Multiple Protein Sequences as a Data Frame}
\usage{
generate_multi_protein_sequences_as_df(num_seq, seq_length)
}
\arguments{
\item{num_seq}{the number of Protein sequences to generate}

\item{seq_length}{the length of each Protein sequence}
}
\value{
a data frame with two columns: "Header" and "Sequence"
}
\description{
This function generates multiple random Protein sequences of a specified length and returns them as a data frame.
}
\examples{
generate_multi_protein_sequences_as_df(5, 50)

}
