% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multi_protein_sequences_as_list.R
\name{generate_multi_protein_sequences_as_list}
\alias{generate_multi_protein_sequences_as_list}
\title{Generate multiple protein sequences as a list}
\usage{
generate_multi_protein_sequences_as_list(num_seq, seq_length)
}
\arguments{
\item{num_seq}{The number of protein sequences to generate.}

\item{seq_length}{The length of each protein sequence.}
}
\value{
A list of protein sequences.
}
\description{
Generate multiple protein sequences as a list
}
\examples{
generate_multi_protein_sequences_as_list(10, 50)

}
