% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multi_rna_seq.R
\name{generate_multi_rna_sequences}
\alias{generate_multi_rna_sequences}
\title{Generate multiple RNA sequences}
\usage{
generate_multi_rna_sequences(num_seq, seq_length)
}
\arguments{
\item{num_seq}{An integer specifying the number of RNA sequences to generate.}

\item{seq_length}{An integer specifying the length of each RNA sequence.}
}
\value{
A character vector of RNA sequences in FASTA format.
}
\description{
This function generates a specified number of RNA sequences with random bases.
}
\examples{
generate_multi_rna_sequences(3, 10)

}
