% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_dna_sequence.R
\name{generate_random_dna_sequence}
\alias{generate_random_dna_sequence}
\title{Generate Random DNA Sequence}
\usage{
generate_random_dna_sequence(length)
}
\arguments{
\item{length}{The length of the DNA sequence to generate.}
}
\value{
A character string representing a random DNA sequence.
}
\description{
Generates a random DNA sequence of a given length.
}
\examples{
generate_random_dna_sequence(10)

}
