% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_protein_sequence.R
\name{generate_random_protein_sequence}
\alias{generate_random_protein_sequence}
\title{Generate a random protein sequence}
\usage{
generate_random_protein_sequence(length)
}
\arguments{
\item{length}{The length of the protein sequence to be generated.}
}
\value{
A character string representing the generated protein sequence.
}
\description{
This function generates a random protein sequence of a given length using the frequencies of each of the 20 amino acids in the human proteome.
}
\examples{
generate_random_protein_sequence(10)

}
