% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_rna_sequence.R
\name{generate_random_rna_sequence}
\alias{generate_random_rna_sequence}
\title{Generate a random RNA sequence}
\usage{
generate_random_rna_sequence(length)
}
\arguments{
\item{length}{the length of the RNA sequence to be generated}
}
\value{
a character string representing the generated RNA sequence
}
\description{
This function generates a random RNA sequence of a given length.
}
\examples{
generate_random_rna_sequence(10)

}
