% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sanger_fastq_sequences.R
\name{generate_sanger_fastq_sequences}
\alias{generate_sanger_fastq_sequences}
\title{Generate Sanger FASTQ Sequences}
\usage{
generate_sanger_fastq_sequences(num_seq, seq_length)
}
\arguments{
\item{num_seq}{The number of sequences to generate.}

\item{seq_length}{The length of the sequences to generate.}
}
\value{
A character vector of the generated Sanger FASTQ sequences.
}
\description{
This function generates a specified number of Sanger FASTQ sequences with a given length.
}
\examples{
generate_sanger_fastq_sequences(2, 10)

}
