% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{get_genetic_code}
\alias{get_genetic_code}
\title{Get Genetic Code}
\usage{
get_genetic_code(table = 1)
}
\arguments{
\item{table}{Integer NCBI genetic code table index}
}
\value{
A named character vector
}
\description{
Returns a mapping of codons to amino acids.
}
