% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_aa_composition}
\alias{plot_aa_composition}
\title{Plot AA Composition}
\usage{
plot_aa_composition(s)
}
\arguments{
\item{s}{A character string containing the protein sequence}
}
\value{
A ggplot object
}
\description{
Visualizes the amino acid composition categorized by biochemical properties.
}
\examples{
prot <- "MKFLVLALAL"
plot_aa_composition(prot)
}
