% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dotplot}
\alias{plot_dotplot}
\title{Plot Dot Plot}
\usage{
plot_dotplot(s1, s2, window = 1)
}
\arguments{
\item{s1}{First sequence}

\item{s2}{Second sequence}

\item{window}{Integer word size for matching (default: 1)}
}
\value{
A ggplot object
}
\description{
Generates a dot plot comparison of two sequences.
}
\examples{
s1 <- "ATGCATGCATGC"
s2 <- "ATGCGTGCATGC"
plot_dotplot(s1, s2, window = 3)
}
