% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gc_skew}
\alias{plot_gc_skew}
\title{Plot GC Skew}
\usage{
plot_gc_skew(s, window = 100)
}
\arguments{
\item{s}{A character string containing the DNA sequence}

\item{window}{Integer window size (default: 100)}
}
\value{
A ggplot object
}
\description{
Generates a sliding window plot of GC skew (G-C)/(G+C).
}
\examples{
data(sars_fragment)
plot_gc_skew(sars_fragment, window = 100)
}
