% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hydrophobicity}
\alias{plot_hydrophobicity}
\title{Plot Hydrophobicity}
\usage{
plot_hydrophobicity(s, window = 9)
}
\arguments{
\item{s}{A character string containing the protein sequence}

\item{window}{Integer window size (default: 9)}
}
\value{
A ggplot object
}
\description{
Generates a sliding window plot of protein hydrophobicity using the Kyte-Doolittle scale.
}
\examples{
prot <- "MKFLVLALAL"
plot_hydrophobicity(prot, window = 3)
}
