% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta_df.R
\name{read.fasta_to_df}
\alias{read.fasta_to_df}
\title{Read a fasta file into a dataframe and assign to the environment}
\usage{
read.fasta_to_df(fasta_file)
}
\arguments{
\item{fasta_file}{The path to the fasta file to be read.}
}
\value{
This function does not return anything. It assigns the resulting dataframe
to the environment.
}
\description{
This function reads a fasta file and creates a dataframe with two columns:
Header and Sequence. The dataframe is then assigned to the environment with the
name same as the fasta file name but without the .fasta extension.
}
\examples{

sample_file_path <- system.file("extdata", "sample_fa.fasta", package = "baseq")

read.fasta_to_df(sample_file_path)

}
