% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta_list.R
\name{read.fasta_to_list}
\alias{read.fasta_to_list}
\title{Read a fasta file into a list and assign to the environment}
\usage{
read.fasta_to_list(fasta_file)
}
\arguments{
\item{fasta_file}{The path to the fasta file to be read.}
}
\value{
This function does not return anything. It assigns the resulting list
to the environment.
}
\description{
This function reads a fasta file and creates a list with two columns:
Header and Sequence. The list is then assigned to the environment with the
name same as the fasta file name but without the .fasta extension.
}
\examples{

sample_file_path <- system.file("extdata", "sample_fa.fasta", package = "baseq")

read.fasta_to_list(sample_file_path)

# Access a specific sequence by name
sample_fa[["sample_seq.1"]]

}
