% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq_df.R
\name{read.fastq_to_df}
\alias{read.fastq_to_df}
\title{Read a Fastq file and store it as a dataframe}
\usage{
read.fastq_to_df(fastq_file)
}
\arguments{
\item{fastq_file}{A character string specifying the path to the Fastq file to be read.}
}
\value{
This function returns a dataframe with three columns: Header, Sequence, and QualityScore.
}
\description{
This function reads a Fastq file and stores it as a dataframe with three columns: Header, Sequence, and QualityScore.
}
\examples{

sample_file_path_two <- system.file("extdata", "sample_fq.fastq", package = "baseq")

read.fastq_to_df(sample_file_path_two)

}
