% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_seq}
\alias{read_seq}
\title{Universal Sequence Reader}
\usage{
read_seq(file, format = "df")
}
\arguments{
\item{file}{Path to the input sequence file}

\item{format}{A string indicating "df" (dataframe) or "list" (default: "df")}
}
\value{
A dataframe or list of the sequence data.
}
\description{
Reads a FASTA or FASTQ file and returns it as a dataframe or list.
}
