% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{rev_comp}
\alias{rev_comp}
\title{Universal Reverse Complement}
\usage{
rev_comp(sequence)
}
\arguments{
\item{sequence}{A character string containing the sequence}
}
\value{
A character string of the reverse complement
}
\description{
Generates the reverse complement of a DNA or RNA sequence.
}
