% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{reverse_translate}
\alias{reverse_translate}
\title{Reverse Translation}
\usage{
reverse_translate(s)
}
\arguments{
\item{s}{A character string containing the protein sequence}
}
\value{
A character string of the resulting DNA sequence
}
\description{
Converts a protein sequence back into DNA using common codons.
}
