% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{rna_to_dna}
\alias{rna_to_dna}
\title{RNA to DNA}
\usage{
rna_to_dna(s)
}
\arguments{
\item{s}{A character string containing the RNA sequence}
}
\value{
A character string of the DNA sequence
}
\description{
Reverse transcribes an RNA sequence into DNA.
}
