% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{rna_to_protein}
\alias{rna_to_protein}
\title{Translate RNA to Protein}
\usage{
rna_to_protein(s, table = 1)
}
\arguments{
\item{s}{A character string containing the RNA sequence}

\item{table}{Integer indicating the NCBI genetic code table (default: 1)}
}
\value{
A list of translated protein sequences
}
\description{
Translates an RNA sequence into protein in all 6 reading frames.
}
