% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sars_fragment}
\alias{sars_fragment}
\title{SARS-CoV-2 Genome Fragment}
\format{
A character string.
}
\source{
NCBI GenBank
}
\usage{
sars_fragment
}
\description{
A small fragment of the SARS-CoV-2 genome used for examples and testing.
}
\keyword{datasets}
