% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{search_motif}
\alias{search_motif}
\title{Motif Searching}
\usage{
search_motif(s, p)
}
\arguments{
\item{s}{A character string containing the sequence}

\item{p}{A character string containing the motif (regex)}
}
\value{
A dataframe with the Start, End, and Match string
}
\description{
Finds all occurrences of a motif in a sequence.
}
