% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{seq_translate}
\alias{seq_translate}
\title{Translation of a DNA sequence}
\usage{
seq_translate(sequence)
}
\arguments{
\item{sequence}{A character string representing a DNA sequence.}
}
\value{
A list of character strings representing the translated protein sequences in all six frames.
}
\description{
This function takes a DNA sequence as input and translates it in all six reading frames.
}
\examples{
sequence <- "ATCGAGCTAGCTAGCTAGCTAGCT"
seq_translate(sequence)
# Returns a list containing the translated protein sequences in all six frames:
# $Frame F1
# [1] "SASRSLE"
# $Frame F2
# [1] "EQGRC*"
# $Frame F3
# [1] "GACT*"
# $Frame R1
# [1] "YLSLSLHLEI"
# $Frame R2
# [1] "STNISLLIP"
# $Frame R3
# [1] "GLGA*"
}
