% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{shuffle_sequence}
\alias{shuffle_sequence}
\title{Shuffle Sequence}
\usage{
shuffle_sequence(s)
}
\arguments{
\item{s}{A character string containing the sequence}
}
\value{
A character string of the shuffled sequence
}
\description{
Randomly permutes the characters of a sequence.
}
