% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlab.R
\name{simulate_digestion}
\alias{simulate_digestion}
\title{Virtual Digestion}
\usage{
simulate_digestion(s, p)
}
\arguments{
\item{s}{A character string containing the DNA sequence}

\item{p}{A character string containing the restriction site (regex)}
}
\value{
A numeric vector of fragment lengths
}
\description{
Simulates restriction enzyme digestion.
}
