% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{simulate_fasta}
\alias{simulate_fasta}
\title{Simulate FASTA File}
\usage{
simulate_fasta(n_seq = 5, seq_len = 100, gc = NULL, type = "DNA", file = NULL)
}
\arguments{
\item{n_seq}{Number of sequences}

\item{seq_len}{Length of each sequence}

\item{gc}{Target GC content}

\item{type}{"DNA" or "RNA"}

\item{file}{Optional file path to save}
}
\value{
A dataframe of simulated sequences
}
\description{
Generates a dummy FASTA dataset.
}
