% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{simulate_fastq}
\alias{simulate_fastq}
\title{Simulate FASTQ File}
\usage{
simulate_fastq(
  n_reads = 5,
  read_len = 100,
  gc = NULL,
  type = "DNA",
  file = NULL
)
}
\arguments{
\item{n_reads}{Number of reads}

\item{read_len}{Length of each read}

\item{gc}{Target GC content}

\item{type}{"DNA" or "RNA"}

\item{file}{Optional file path to save}
}
\value{
A dataframe of simulated reads
}
\description{
Generates a dummy FASTQ dataset.
}
