% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlab.R
\name{simulate_pcr}
\alias{simulate_pcr}
\title{PCR Simulator}
\usage{
simulate_pcr(template, fwd, rev_p)
}
\arguments{
\item{template}{A character string containing the DNA template}

\item{fwd}{A character string of the forward primer}

\item{rev_p}{A character string of the reverse primer}
}
\value{
A numeric vector of amplicon sizes
}
\description{
Simulates a PCR reaction and predicts amplicon sizes.
}
