% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{simulate_sequence}
\alias{simulate_sequence}
\title{Simulate Sequence}
\usage{
simulate_sequence(len, gc = NULL, type = "DNA")
}
\arguments{
\item{len}{Integer length of the sequence}

\item{gc}{Numeric target GC content (0 to 1)}

\item{type}{"DNA" or "RNA"}
}
\value{
A character string of the simulated sequence
}
\description{
Generates a random DNA or RNA sequence.
}
