% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarize_fasta}
\alias{summarize_fasta}
\title{FASTA Summary}
\usage{
summarize_fasta(file)
}
\arguments{
\item{file}{Path to the FASTA file}
}
\value{
A summary dataframe
}
\description{
Generates a comprehensive summary of a multi-FASTA file.
}
\examples{
\donttest{
# summarize_fasta("path/to/my.fasta")
}
}
