% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{translate}
\alias{translate}
\title{Generic Translate}
\usage{
translate(x, ...)
}
\arguments{
\item{x}{A baseq_dna or baseq_rna object}

\item{...}{Additional arguments}
}
\value{
A list of translated sequences
}
\description{
Generic function to translate DNA or RNA to protein.
}
