% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df_to_fastq.R
\name{write.df_to_fastq}
\alias{write.df_to_fastq}
\title{Write a FASTQ file from a dataframe of reads}
\usage{
write.df_to_fastq(df)
}
\arguments{
\item{df}{A dataframe containing reads in the format "Header", "Sequence", and "QualityScore".}
}
\value{
A FASTQ file with the same name as the input dataframe.
}
\description{
Write a FASTQ file from a dataframe of reads
}
\examples{
sample_file_path_two <- system.file("extdata", "sample_fq.fastq", package = "baseq")
read.fastq_to_df(sample_file_path_two)
write.df_to_fastq(sample_fq)

}
