% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R, R/write_fastq.R
\name{write.fasta}
\alias{write.fasta}
\alias{write_fasta}
\title{Write fasta file}
\usage{
write.fasta(sequences, output_dir = getwd())

write_fasta(sequences, output_dir = getwd())

write_fasta(sequences, output_dir = getwd())
}
\arguments{
\item{sequences}{A data frame or a list of sequences to be written to a fasta file.}

\item{output_dir}{The directory where the output fasta file will be saved. Default is the current working directory.}
}
\description{
This function writes a fasta file from either a data frame or a list of sequences.
}
\details{
If the input object is a data frame, the function expects the data frame to have two columns named "Header" and "Sequence". The function will write each record to the fasta file.

If the input object is a list, the function will loop over the list elements and write each sequence to the file in FASTA format in the format.
}
\examples{
# Write fasta file from a data frame
# df <- data.frame(Header = c("seq1", "seq2"), Sequence = c("ATCGATCG", "CGATCGAT"))
# write_fasta(df)

# Write fasta file from a list
# seq_list <- list("ATCGATCG", "CGATCGAT")
# write_fasta(seq_list)

}
