% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fastq.R
\name{write.fastq}
\alias{write.fastq}
\title{Write dataframe or list to FASTQ file}
\usage{
write.fastq(input, output_dir = getwd())
}
\arguments{
\item{input}{A dataframe or list containing sequence data}

\item{output_dir}{A string specifying the directory where the output file will be written (default is the current working directory)}
}
\value{
This function does not return anything, but writes the input to a FASTQ file
}
\description{
This function takes as input a dataframe or list and writes it to a FASTQ file format.
}
\examples{
# Write a dataframe to a FASTQ file
# write.fastq(df, "output_directory")

# Write a list to a FASTQ file
# write.fastq(sequence_list, "output_directory")

}
