% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_seq}
\alias{write_seq}
\title{Universal Sequence Writer}
\usage{
write_seq(x, file)
}
\arguments{
\item{x}{A sequence object (dataframe or list)}

\item{file}{Path to the output sequence file}
}
\value{
Invisible TRUE
}
\description{
Writes a sequence object (dataframe or list) to a FASTA or FASTQ file.
}
