% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_RNA.R
\name{clean_RNA_file}
\alias{clean_RNA_file}
\title{Clean RNA file}
\usage{
clean_RNA_file(input_file)
}
\arguments{
\item{input_file}{The name of the input multi FASTA file.}
}
\value{
A character string specifying the path to the output FASTA file.
}
\description{
This function reads a multi FASTA file containing RNA sequences, removes any characters other
than A, T, G, and C, and writes the cleaned sequences to a new multi FASTA file. The output
file name is generated from the input file name with the suffix '_clean.fasta'.
}
\examples{
sample_file_path_three <- system.file("extdata", "sample2_fa.fasta", package = "baseq")
clean_RNA_file(sample_file_path_three)

}
