% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Transcription of a DNA sequence}
\usage{
dna_to_rna(sequence)
}
\arguments{
\item{sequence}{A character string representing a DNA sequence.}
}
\value{
A character string representing the RNA transcript of the input DNA sequence.
}
\description{
This function takes a DNA sequence as input and returns its RNA transcript.
}
\examples{
sequence <- "ATCGAGCTAGCTAGCTAGCTAGCT"
dna_to_rna(sequence)
# Returns "AUCGAGCUAGCUAGCUAGCUAGCU"
}
