% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_content.R
\name{gc_content}
\alias{gc_content}
\title{Calculate GC content of a DNA sequence}
\usage{
gc_content(sequence)
}
\arguments{
\item{sequence}{A character string containing the DNA sequence.}
}
\value{
A numeric value representing the percentage of nucleotides in the sequence that are G or C.
}
\description{
Calculates the percentage of nucleotides in a DNA sequence that are either guanine (G) or cytosine (C).
}
\examples{
sequence <- "ATCGAGCTAGCTAGCTAGCTAGCT"
gc_content(sequence)
50
}
