% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read.fasta}
\alias{read.fasta}
\alias{read_fasta}
\title{Read a FASTA file as a dictionary}
\usage{
read.fasta(file_path)

read_fasta(file_path)
}
\arguments{
\item{file_path}{The path to the FASTA file.}
}
\value{
A list where each key is a sequence identifier and the corresponding
value is the sequence itself.
}
\description{
This function reads a FASTA file as a character vector and returns a list
where each key is a sequence identifier and the corresponding value is the
sequence itself.
}
\examples{
# fasta_dict <- read_fasta("path/to/your/fasta/file.fasta")
# seq1 <- fasta_dict[["sequence1"]]

}
